Fixed errors exposed through testing in a Rails 3 app
The test cases don't provide full coverage. Test cases that test
the operation inside a real ActionView doing templating are needed.
The problem was use of the output_buffer, this is used with haml and erb
in a way that I don't fully understand, bottom line being don't assign to it,
it messes up rendering of your views. Used form_tag_in_block for guidance on how to
provide a view helper. Refactored tests to use blocks that output html_safe strings
similar to what haml or erb would give.